better timeout checking for dailycensor

+++ extensions/dailycensor.py
@@ -2,6 +2,7 @@
 import asyncio
 import discord
 import re
+from datetime import datetime, timedelta
 from discord.ext import commands
 
 from common import *
@@ -31,15 +32,39 @@ class DailyCensor(commands.Cog):
         channel = message.channel
         role = guild.get_role(THIN_ICE_ROLES[guild_key])
         if role in author.roles:
-            await message.delete()
-            await channel.send(
-                "oopsie woopsie {} did a fuckie wuckie!".format(author.mention)
-            )
-            return
+            start_time = datetime.now() - timedelta(days=1)
+
+            def action_added_thin_ice_role(action):
+                if action.created_at < start_time:
+                    return False
+                if action.target != author:
+                    print("not the right author {}".format(author.display_name))
+                    return False
+                print("the right author {}".format(author.display_name))
+                if role in action.before.roles:
+                    print("had the role before")
+                    return False
+                print("did not have the role before")
+                if role not in action.after.roles:
+                    print("did not have the role after")
+                    return False
+                print("had the role after")
+                return True
+
+            last_swear = await guild.audit_logs(
+                action=discord.AuditLogAction.member_role_update,
+                after=start_time,
+                limit=None,
+            ).find(action_added_thin_ice_role)
+            if last_swear != None:
+                await message.delete()
+                await channel.send(
+                    "oopsie woopsie {} did a fuckie wuckie!".format(author.mention)
+                )
+                return
+            await author.remove_roles(role)
         await channel.send("That's your one cuss for the next 24 hours.")
         await author.add_roles(role)
-        await asyncio.sleep(86400)
-        await author.remove_roles(role)
 
 
 def setup(bot):
